# tauri/packages/types/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Springtale management API",
    "description": "Local-first bot automation. Binds loopback; bearer-authenticated.",
    "license": {
      "name": "AGPL-3.0"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/templates/{name}": {
      "post": {
        "tags": [
          "templates"
        ],
        "summary": "POST /templates/{name} — write the named template.",
        "description": "The daemon picks the destination directory and returns the path in\nthe response. No `dir` field accepted — the caller cannot influence\nwhere files land (OWASP ASVS §12.3).",
        "operationId": "templates_write",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Template name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Files written",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WriteReport"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "WriteReport": {
        "type": "object",
        "description": "Outcome of writing a template to disk.",
        "required": [
          "template",
          "dir",
          "created"
        ],
        "properties": {
          "created": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "dir": {
            "type": "string"
          },
          "template": {
            "type": "string"
          }
        }
      }
    }
  },
...
}